% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amoebiasis}
\alias{amoebiasis}
\title{Amoebiasis Epidemiologic Data From Monthly Surveillance Data}
\format{A data frame of 23244 rows and 5 variables:
\itemize{
   \item \code{province}: province name
   \item \code{year}: year
   \item \code{month}: month
   \item \code{incidence}: number of new cases
   \item \code{mortality}: number of death cases
}}
\source{
data provide by the Vietnamese General Department of Preventive
Medicine (GDPM) and by the National Institute of Hygiene and Epidemiology
(NIHE), and supported by the Oxford University Clinical Research Unit (OUCRU)
and the French National Research Institute for Sustainable Development (IRD).
}
\usage{
data(amoebiasis)
}
\description{
A dataset containing \code{incidence} and \code{mortality} data for
amoebiasys from the Vietnamese GDPM, aggregated by month and province from
January 1980 to December 2015.
}
\details{
The incidence corresponds to the number of new cases and the mortality
corresponds to the number of deaths. Before 1991, \code{NA} corresponds to
"no case" or "no report". After 1991, \code{NA} corresponds to "no report"
and \code{0} to "no case".
In 1990, the number and the name of the province are different for amoebiasis
and hepatitis compared to the other disease. In 1990, the number and name of
the province changed in Vietnam, but some hospital made only the change a
year after.
}
\keyword{datasets}

