% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdpm.R
\docType{package}
\name{gdpm}
\alias{gdpm}
\alias{gdpm-package}
\title{gdpm : Surveillance Data from the Vietnamese GDPM}
\description{
Monthly surveillance data aggregated by province by the Vietnamese General
Department of Preventive Medicine (GDPM)
}
