% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_map.R
\name{gdpm_choropleth}
\alias{gdpm_choropleth}
\title{Makes a spatio-temporal choropleth map of a disease}
\usage{
gdpm_choropleth(df, ye, mo, x, y, sel_value = "incidence", n = 6,
  col = heat.colors(6), style = "quantile", col_na = "grey",
  fixedBreaks = NULL, locate = FALSE, pos = "top-left", distrib = TRUE,
  n_round = 0, pos_brks = TRUE, postext = "left", h = 0.75, w = 0.75,
  tl = 0.2, s = 0.2, ...)
}
\arguments{
\item{df}{a data frame containing the epidemiological data (e.g. \code{ili})
and at least the colums "year", "month", "province" and "incidence" and/or
"mortality"}

\item{ye}{an year to indicate the temporal information}

\item{mo}{month to precise the temporal information}

\item{x}{a value for the x coordinate of the top-left point of the legend}

\item{y}{a value for the y coordinate of the top-left point of the legend}

\item{sel_value}{a character value which can be "incidence" (by default) or
"mortality" used to select the data expressed in the map}

\item{n}{a numeric indicating the number of intervals to represent the data
(by default, \code{n = 6})}

\item{col}{a vector of colors to use for the map (by default,
\code{col = heat.colors(6)}.The colors from the package RColorBrewer can also
 be used.}

\item{style}{a character value issued from the \code{classint} package, and
used to select a method for the different way of calculating the intervals
(by default \code{style = "quantile"})}

\item{col_na}{the color with which to represent the missing values
(by default \code{col_na = "grey"})}

\item{fixedBreaks}{issued from the \code{classint} package. By default
\code{NULL} but if a vector value is inputed, it will be used to specifen the
 breaks}

\item{locate}{if TRUE, call the function \code{locator} to indicate the
top-left point of the legend}

\item{pos}{by default \code{top-left}, but can be \code{top-right,
bottom-left or bottom-right}, used to indicate the position of the data
if \code{x, y} are not indicated}

\item{distrib}{if TRUE, print on the map, the distribution of the values by
intervals}

\item{n_round}{integer indicating the number of significant digits to be used
in the legend, by default (\code{0})}

\item{pos_brks}{if TRUE, the breaks values will all be positive, the first
break will be superior or equal to zero, by default (\code{TRUE}). If false,
allows negative value for breaks}

\item{postext}{define the side of the legend text, by default \code{left}
but can be \code{right}}

\item{h}{legend parameter expressing the height of one rectangle
in the legend}

\item{w}{legend parameter expressing the width of the legend}

\item{tl}{legend parameter expressing the length of the tick}

\item{s}{legend parameter expressing the space between the text and the
tick}

\item{...}{if need to imput more text parameters for the legend}
}
\description{
Makes a spatio-temporal choropleth map of a disease
}
\details{
The \code{n} parameter can be overuled by the number of breaks
obtain by certain style. By default, the number of intervall will be decide
by the \code{n} parameter but for certain \code{style} like "pretty", this
value can be overruled. Please for more information, look at the
documentation of the classint package.
\cr\cr The map selected is automatically the one corresponding to the first
year of the dataset inputed.
}
\examples{
library(gdpm)
ili <- getid_("ili", from = 2004, to = 2004)
# A choroplet map of the ILI data:
gdpm_choropleth(ili, 2004, "April")

# with some data transformations in order to reflect better the contrasts:
gdpm_choropleth(ili, 2004, "April", x = 93, y = 18, n = 6,
col = heat.colors(6), style = "jenks")

# using some other color palettes, for examples the ones from the
# RColorBrewer package or the one in the color Palettes of R :
library(RColorBrewer)
# to see the available color palettes:
display.brewer.all()
gdpm_choropleth(ili, 2004, "April", x = 93, y = 18, n = 6,
col = heat.colors(6), style = "jenks")
gdpm_choropleth(ili, 2004, "April", x = 93, y = 18, n = 6,
 col = "YlOrBr", style = "jenks")

# changing the color of the missing values:
ili80 <- getid(ili, from = 1980, to = 1980)
gdpm_choropleth(ili80, 1980, "April", col_na = "chartreuse")

# changing the legend text parameters
gdpm_choropleth(ili80, 1980, "April", n = 6, col = heat.colors(6),
style = "jenks", cex = 0.5)
# Print the numeric legend with 2 decimals
gdpm_choropleth(ili80, 1980, "April", n = 6, n_round = 2)

gdpm_choropleth(ili80, 1980, "April", n = 6, distrib = FALSE,
  col = heat.colors(6), style = "jenks")

# By default, the legend is on the top left of the figure, but the position
# can be easily change by using the parameters pos
 # top left
gdpm_choropleth(ili, 2004, "April", n = 6, col = heat.colors(6),
  style = "jenks")
gdpm_choropleth(ili, 2004, "April", n = 6, pos = "top-left",
  col = heat.colors(6), style = "jenks")
# top right
gdpm_choropleth(ili, 2004, "April", n = 6, pos = "top-right",
  col = heat.colors(6), style = "jenks", distrib = FALSE)
# bottom left
gdpm_choropleth(ili, 2004, "April", n = 6, pos = "bottom-left",
  col = heat.colors(6), style = "jenks")
# bottom right
gdpm_choropleth(ili, 2004, "April", n = 6, pos = "bottom-right",
  col = heat.colors(6), style = "jenks", distrib = FALSE)

# By default, the function print the map of the incidence value for one
# month of one year, but the mortality can also be printed.
gdpm_choropleth(ili, 2004, "April", n = 6, sel = "incidence",
  col = heat.colors(6), style = "jenks")
gdpm_choropleth(ili, 2004, "January", n = 3, sel = "mortality",
  col = heat.colors(3), style = "jenks")

# The intervals used to expressed the value can be input directly as
# parameters in the function via the parameters fixedBreaks.
# The given breaks should be of length n+1.
gdpm_choropleth(ili, 2004, "April", col = heat.colors(3),
  fixedBreaks = c(0, 5000, 10000, 15000))

# Using the locator to choose where to print the legend
\dontrun{
 gdpm_choropleth(ili80, 1980, "April", n = 6, locate = TRUE,
  col = heat.colors(6), style = "jenks")
gdpm_choropleth(ili80, 1980, "April", n = 6, locate = TRUE, x = 95, y = 34,
  col = heat.colors(6), style = "jenks")
}

}
