% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging_function.R
\name{merging}
\alias{merging}
\title{Merges provinces}
\usage{
merging(df, from, to)
}
\arguments{
\item{df}{An epidemiological data frame (e.g. \code{ili} or \code{dengue}).
Should contain at least the variables \code{province}, \code{year},
\code{month} and the variables \code{incidence} and \code{mortality}.}

\item{from}{Initial date of the time range, of the class \code{Date}.}

\item{to}{Final date of the data, of the class \code{Date}.}
}
\value{
A object of the same class as \code{df} in which all the provinces
that needed to be merged (according to the time range) are merged.
}
\description{
Merges epidemiological data accordingly to a time range and merges the
provinces concern by a split/combined event if necessary and returns a data
frame for the time range imputed.
}
\examples{
# For a time range starting on the first of January 1980, finishing on the
# first of January 2004, to obtain a data frame with all the provinces and in
# which all the provinces that needed to be merged, are merged:
library(gdpm)
data("dengue")
data("hepatitis")
data("amoebiasis")
range(dengue$year)
merging(dengue,"1990-01-01","2004-01-01")
range(hepatitis$year)
merging(hepatitis,"1980-01-01","2009-01-01")
range(amoebiasis$year)
merging(amoebiasis,"1980-01-01","2009-01-01")
}

