% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diseases}
\alias{diseases}
\title{Summary Data From Monthly Surveillance Data}
\format{A data frame of 30 rows and 5 variables:
\itemize{
   \item \code{file}: name of data frame
   \item \code{nobs}: number of observations
   \item \code{nvar}: number of variables
   \item \code{year_start}: first year of the dataset
   \item \code{year_end}: last year of the dataset
}}
\source{
data provide by the Vietnamese General Department of Preventive
Medicine (GDPM) and by the National Institute of Hygiene and Epidemiology
(NIHE), and supported by the Oxford University Clinical Research Unit (OUCRU)
and the French National Research Institute for Sustainable Development (IRD).
}
\usage{
data(diseases)
}
\description{
A dataset containing the description of all the data frames contained in the
package.
}
\details{
The dataset contains, the names of each data frame, the time range of the
data. This table can be used as a resume of all the other data frame contain
in this package and may be useful for computing on the various files.
}
\keyword{datasets}
