% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging_function.R
\name{getid_}
\alias{getid}
\alias{getid_}
\title{Gets Infectious Diseases}
\usage{
getid_(disease, from = "2004-01-01", to = "2015-12-31")

getid(disease, from = `2004-01-01`, to = `2015-12-31`)
}
\arguments{
\item{from}{Initial date of the time range, of the class \code{Date}.}

\item{to}{Final date of the data, of the class \code{Date}.}

\item{df}{An epidemiological data frame (e.g. \code{ili} or \code{dengue}).
Should contain at least the variables \code{province}, \code{year},
\code{month} and the variables \code{incidence} and \code{mortality}.}
}
\value{
A object of the same class as \code{df} in which all the provinces
that needed to be merged (according to the time range) are merged.
}
\description{
Merges epidemiological data accordingly to a time range and merges the
provinces concern by a split/combined event if necessary and returns a data
frame for the time range imputed.
}
\details{
A dataset called \code{diseases} contains the description of all the
epidemiological data frames available in this package. The dataset is ordered
by the names of each disease and the time range of the data.
This table can be used as a resume.
}
\examples{
library(gdpm)
# Load a resume table of all the epidemiological data frame contains in the
# package.
diseases
# Return a data frame in which all the provinces that needed to be merged
# (according to the time range) are merged.
getid_("dengue", "1990-01-01", "2004-01-01")
getid(hepatitis, `1980-01-01` ,`2009-01-01`)
# By default, the time range selected is from 2004-01-01 to 2015-12-31.
getid_("cholera")
getid(hepatitis)
# For the time range, just the year can be entered as parameters and by
# default, the first day of the year "from" and the last day of the year
# "to" will be selected as time range.
getid_("dengue", "1990", "2004")
getid(hepatitis,1980 ,2009)
}

